Move ProfileContext to AIF
